Implement controller swap and new Windows 11 interface (#110)
* initial commit

* implement Windows 11 ModernWpf (WIP)

* more work on controller swap (wip)

* About page with Windows 11 style

* Overlay page with Windows 11 style

* Settings page with Windows 11 style

* more work on controller swap. migrated settings to ControllerService.dll.config

* fix crash on settings pull from client

* rely on PipeServerSettings to pick UserIndex rather than PipeServerHandheld

* code cleanup

* fix all ToggleSwitch orientation (at the cost of displaying text)

* replace ComboBox with RadioButtons

* properly display trackpad position on overlay preview

* (re)order back and start button on overlay settings

* (re)order Left and Right joystick on Profile page

* center cB_Buttons

* change UMC glyph on profile page

* add missing trigger on UMC expander click

* fix OverlayTrigger_SelectionChanged

* build 0.9.6.1

* Remove temp code to highlight vibration motors in 3D overlay.

* implement trackpads size slider

* implement GetCurrentReading(ratio)

* update profile naming

* implement Accelerations and AngularVelocities with XInputSensorFlags

* pick currently used profile when foreground process has changed

* rename microsoftLogger to logger

* send toast notification when input device is changed

* temporary fix to update manager

+++ ControllerCommon/Devices/Device.cs
@@ -1,9 +1,4 @@
 ﻿using ControllerCommon.Utils;
-using HidSharp;
-using System.Collections.Generic;
-using System.Linq;
-using System.Management;
-using System.Runtime.InteropServices;
 using Windows.Devices.Sensors;
 using static ControllerCommon.Utils.DeviceUtils;
 
@@ -28,7 +23,7 @@ namespace ControllerCommon.Devices
         protected Inclinometer inclinometer;
         public bool hasInclinometer;
 
-        public double WidthHeightRatio = 1.0d;
+        public float WidthHeightRatio = 1.0f;
 
         protected Device(string ManufacturerName, string ProductName)
         {
